requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask_jwt_extended==4.5.3
python-dotenv==1.0.0
PyYAML==6.0.1
pyodbc==5.1.0
requests~=2.31.0
pandas~=2.2.2